+++ CHANGELOG.md
@@ -3,11 +3,15 @@
 ### Added
 - `HttpCacheManager.createLazyStream()` - Creates a lazy-loading cache stream that automatically manages `HttpCacheStream` lifecycle
   - Streams are created on-demand when first requested
-  - Automatically disposed after `autoDisposeDelay` when no active requests remain
-  - Reduces resource usage for intermittently-accessed URLs
-  - Accepts optional `file`, `config`, and `autoDisposeDelay` parameters
+  - Automatically disposed after `StreamCacheConfig.autoDisposeDelay` when no active requests remain
+  - Supports redirects when using `LazyCacheStream.requestHeader` to make requests (e.g., m3u8 files).
+  - Accepts optional `file`, and `config`, parameters
 
 * Breaking: Removed cacheDir and customHttpClient parameters in HttpCacheManager.init(). To specify a custom cache directory and/or http client, provide a GlobalCacheConfig during initalization. (See GlobalCacheConfig.init)
+
+* Breaking: Removed autoDisposeDelay parameter in `HttpCacheManager.createServer()`. Use `StreamCacheConfig.autoDisposeDelay` instead.
+
+* Added an optional 'cacheFileResolver' to GlobalCacheConfig. When provided, the resolver is used by the manager to generate cache file paths instead of the default path generation
 
 * Improved cache I/O write performance by immediately writing data to file system
 
@@ -16,6 +20,8 @@
 * Added an optional port parameter to HttpCacheManager.init and HttpCacheManager.createServer to specify the cache server port instead of using a random available port.
 
 * Adjust default minChunkSize to 128KB and rangeRequestSplitThreshold to 5MB for improved cache stream performance and responsiveness.
+
+* Update example project with HLS video using LazyCacheStream
 
 
 ## 0.0.3